Merge initializer and environment into a single function
Intead of having two ways how an environment can be initialized, merge the functionality of the two functions into `inEnvironment`. The function either accepts an object hash defining the environment or a environmentProvider that builds up the environment on the worker task. The environment provider can accept a unlimited number of arguments.