Restore Worker when not all functions could be retrieved
Up to now, the browser worker thread thrown an error if a function request could not be fulfilled. This causes several issues
1. The parallel promise is not rejected... instead it is never fulfilled
2. The worker thread stays in the wait for function response state forever

This change fixes the issue by always returning a function response but explicitly list the missing functions. The worker then changes back to idle state and rejects the work (what results in a rejected promise)