Reduce Memory Requirements
The memory requirements in the worst Case are still O(2N) since all sub results are copied into the results array. However, as long as the tasks almost complete sequentially, the memory consumption should be far less than 2n. This is achieved by sequentially joining the sub results instead of joining the end results in one step. If the results arrive strictly sequentially, than the memory requirements are O(N+1).